requirements.txt
attrs==20.1.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Flask==2.2.5
google-api-core==2.15.0
google-auth==2.25.2
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
idna==3.6
importlib-metadata==6.7.0
iniconfig==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
more-itertools==8.5.0
packaging==20.4
Pillow==7.2.0
pluggy==0.13.1
protobuf==4.24.4
py==1.9.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==2.4.7
pytest==6.0.1
requests==2.31.0
rsa==4.9
six==1.15.0
toml==0.10.1
typing_extensions==4.7.1
urllib3==2.0.7
Werkzeug==2.2.3
zipp==3.1.0
